move MomentumIterativeMethod into a separate file
